testing box reading
most mp4 boxes have 32 bit headers, but some have 64 bit headers. currently unsure how to distinguish both. the iso part 12 documentation says that the size in the box header would be 1 if it has a 64 bit header, but it is unclear what exactly they are referring to in binary terms.