referencing data array function
contemplating design of the parser going forward.
copying each MPEG-4 box field into a new array would be convenient if I put a terminating char at the end of each array. but this takes double the space (+ terminating char). I guess original object can be freed once parsed.
The other approach is to just reference the data from the original array using a pointer. This would require me to manage how many bytes to read for everything.
The parser is not at the level where I can see the direct impacts of this decision yet which is why I have been thinking about it.
I am leaning towards copying the data and adding a terminating char, then freeing the original box data. In that case I should look into generalizing my linkedlist implementation so far so that I can replace normal box structs with other more specific box structs. I would consider the other option more  (of using a pointer to each start of data field in the data array) if I can think of a good way to determine the end of each data field in the char array.
Some ideas that come to mind for that approach are things such as using #define fieldname_size but that seems like a very cumbersome solution. Another idea is checking if the current address of a data field overlaps with any of the starting addresses of the other fields (or just the next field). I think this one could work although it is a question of giving up efficiency vs memory usage at this point.

In other news I've learned about bit fields and the existence of uint8_t, uint16_t, uint32_t, uint64_t, so I might be refactoring the code with these as well as the decision I make about storing data.