linkedList, bitUtility, & separation of concerns
It's kind of hard revisiting a code base that requires a lot of knowledge of a specific topic. I'm back to reading the MPEG-4 docs though. Installed WSL on my machine so I can finally use Makefiles! Linkedlist got it's own c file because it's really messy having it mixed into the mpeg parsing code.

It's a little scary making such major changes to already working code with something that requires precision such as this. A forgotten edgecase while migrating functions or interfaces can be bad. Hopefully I didn't forget anything with the linkedlist migration!

I have found that the bigger the codebase gets the more generalization and abstraction is needed. Separation of concerns is truly great. But a lot of design and convention decisions need to be made.

In closing, I am happy that the project is taking shape. Although I hoped to get the sample boxes parsed in this commit, a lot of maintenance work needed to be done.