char to float not working and types
unable to convert float bitpattern stored in a char into a float. A possible solution is to attempt to read it as a float when opening the file or open the file again and read up to that point. both solutions might not work due to the endianness being reversed hence bit shifting is needed but not allowed for floats.