parsing sample atoms/boxes
Idea:
although it sounds a little messy, maybe for handling missing boxes such as stss, ctts, etc.. and having to make exceptional flows in the main program, instead the searching function can indicate that not found by creating a box of that type with null (or special value?) size/data that can be used in the individual xParseBox() functions to return things based on exceptional flow there.

This does sound nice for code organization, but I am not sure if it is a good idea or not to hide exceptional flow behind some special values instead of explicitly not found NULL return.