real time to media time to sample number
I realized that making all the type structs contain pointers to their data gets a little annoying. There is not point in having that extra dereference and it's just more memory accesses. I am thinking of getting rid of it but that is a huge refactor of the code. Although it might not be so bad if I just turned char to int and unsigned int to return a stack int/unsigned int then edit the structs to reflect this. the issue lies within the char* that would complicate things since some values can be normal stack arrays  such as type (4 bytes), version (1 byte), flags (3 bytes), but other things like boxData would still need to be a pointer to a heap array.