sample number to sample size and sample offset
reusable functions can become hard to design. an example in this commit is the (to be decided) interface for sampleNumberToSampleSize. the function can take the new sampleInfo type as a parameter and put the sample size in it, or it can just return it. doing both feels unnecessary since unused return values don't seem like a good thing. but this function is used in getSampleOffsetInChunk wihtout a sample atom associated with the data request. this is where the problem lies when trying to make the function reusable and give a consistent interface to all the sample finding related functions.

the sampleInfo was added because I didn't like the idea of passing a random variable into sampleNumberToChunkNumber() to get sampleIndexInChunk back. but this changes the interface for the function compared to the other related ones. but I just realized that the sample offset in chunk can be calculated within this function instead and just exclude sampleNumberToSampleSize from the "having sampleInfo passed to it" interface. I guess using commit descriptions to reflect is actually pretty helpful.