refactoring table types
The code feels a lot better when the types are simpler. The changes that needed to be made to the types came to light when I wanted to binary search several of the arrays. With the state of the types at the time, it would have needed a separate comparator function for each table being searched. Although that is not that bad, I didn't particularly like other things about the types such as the verbose names, and the unknown size (have to remember which mpeg data variable was the total entries of the array. I think now the types are much clearer.