LZW 70%~ speed improvement
using gprof, the following were the results for running the make test build:

  %   cumulative   self          total
 time   seconds   seconds    calls  name
 41.89      2.69     2.69 254267383  gif_intToString
 33.85      4.86     2.17  7176942    gif_arrayConcat

I realized after implementing lzw, that the key of the code table does not need to be a string just because "cc" and "eoi" are part of the table. Those could be easily represneted by their codetable index counterpart. This lead to changing the hashmap interface to take in an array as key and u32 as value, instead of a string for both.